refactor: 입력 형식 확인 함수 파일 분리
